+++ src/molecule/CurrentUserWidget.tsx
@@ -1,31 +1,4 @@
-import { useQuery } from '@tanstack/react-query';
-import { fetcher } from '@/hooks/fetcher';
-
-interface UserInfo {
-  user_id: string;
-  nickname: string;
-  provider: string;
-  third_party_id: string;
-  prof_img: string;
-  mmr: number;
-  two_factor_authentication_type?: string;
-  two_factor_authentication_key?: string;
-  created: Date;
-  deleted?: Date;
-}
-
-function useCurrentUser() {
-  const data = useQuery<UserInfo>({
-    queryKey: ['/user'],
-    queryFn: async () => {
-      const res = await fetcher('/user');
-
-      if (res.ok) return res.json();
-      throw res;
-    },
-  });
-  return data;
-}
+import { useCurrentUser } from "@/hooks/user";
 
 function CurrentUserWidget() {
   const { data, isLoading, isError } = useCurrentUser();
